feat: implement analysis chart save functionality and improve code organization
- Add BotIntegrationAnalysisChart model with migration
- Create modern analysis results viewer UI
- Split models into separate files for better maintainability
- Fix role name typos (CHATML_ROLE_ASSISSTANT -> CHATML_ROLE_ASSISTANT)
- Reorganize imports and improve type hints
- remove old View Results UI and merge into single 'View Analytics'
