+++ bots/signals.py
@@ -3,8 +3,9 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 from django.utils import timezone
 
-from bots.models import Message, CHATML_ROLE_ASSISSTANT
+from bots.models import Message
 from bots.tasks import msg_analysis
+from daras_ai_v2.language_model import CHATML_ROLE_ASSISTANT
 
 
 @receiver(post_save, sender=Message)
@@ -16,7 +17,7 @@ def run_after_message_save(instance: Message, **kwargs):
         # analysis is not running
         and not instance._analysis_started
         # this is the assistant's response
-        and instance.role == CHATML_ROLE_ASSISSTANT
+        and instance.role == CHATML_ROLE_ASSISTANT
     ):
 
         @transaction.on_commit
