+++ daras_ai_v2/analysis_results.py
@@ -1,87 +1,138 @@
-import json
+import datetime
+import typing
 from collections import Counter
 from functools import partial
 
 import gooey_gui as gui
-from django.db.models import IntegerChoices
-from gooey_gui import QueryParamsRedirectException
 
-from app_users.models import AppUser
-from bots.models import BotIntegration, Message
+from bots.models import BotIntegration, DataSelection, GraphType, Message
+from daras_ai.image_input import truncate_text_words
 from daras_ai_v2 import icons
-from daras_ai_v2.base import RecipeTabs
-from daras_ai_v2.copy_to_clipboard_button_widget import copy_to_clipboard_button
 from daras_ai_v2.grid_layout_widget import grid_layout
 from daras_ai_v2.workflow_url_input import del_button
-from gooeysite.custom_filters import related_json_field_summary, JSONBExtractPath
+from gooeysite.custom_filters import JSONBExtractPath, related_json_field_summary
 from recipes.BulkRunner import list_view_editor
-from recipes.VideoBots import VideoBotsPage
-from widgets.author import render_author_from_user
+
+if typing.TYPE_CHECKING:
+    import plotly.graph_objects as go
 
 
-class GraphType(IntegerChoices):
-    display_values = 1, "Display as Text"
-    table_count = 2, "Table of Counts"
-    bar_count = 3, "Bar Chart of Counts"
-    pie_count = 4, "Pie Chart of Counts"
-    radar_count = 5, "Radar Plot of Counts"
+# Modern color palette with 27 distinct colors
+COLOR_PALETTE = [
+    "#FF6B6B", "#45B7D1", "#96CEB4", "#FFEEAD", "#D4A5A5", "#9B59B6", "#3498DB", "#E74C3C", "#2ECC71", "#F1C40F",
+    "#E67E22", "#1ABC9C", "#9B59B6", "#34495E", "#E74C3C", "#3498DB", "#2ECC71", "#F1C40F", "#E67E22", "#1ABC9C",
+    "#9B59B6", "#34495E", "#E74C3C", "#3498DB", "#2ECC71", "#F1C40F",
+]  # fmt: skip
 
 
-class DataSelection(IntegerChoices):
-    all = 1, "All Data"
-    last = 2, "Last Analysis"
-    convo_last = 3, "Last Analysis per Conversation"
-
-
-def render_analysis_results_page(
+def render_analysis_results_viewer(
     bi: BotIntegration,
-    current_url: str,
-    current_user: AppUser | None,
-    title: str | None,
-    graphs_json: str | None,
+    start_date: datetime.datetime,
+    end_date: datetime.datetime,
+    label: str = '<i class="fa-solid fa-head-side-brain"></i> Analysis Results',
+    key: str = "analysis_graphs",
 ):
-    render_title_breadcrumb_share(bi, current_url, current_user)
-
-    if title:
-        gui.write(title)
-
-    if graphs_json:
-        graphs = json.loads(graphs_json)
-    else:
-        graphs = []
-
-    _autorefresh_script()
-
     try:
-        results = fetch_analysis_results(bi)
+        results = fetch_analysis_results(bi, start_date, end_date)
     except Message.DoesNotExist:
         results = None
+    with (
+        gui.styled("& .accordion-header { font-weight: 400; font-size: 1.4rem; }"),
+        gui.div(),
+        gui.expander(label),
+    ):
+        selected_graphs = render_editor(bi, results, key)
     if not results:
-        with gui.center():
-            gui.error("No analysis results found")
         return
+    render_graphs(bi, results, selected_graphs)
 
+
+@gui.cache_in_session_state
+def fetch_analysis_results(
+    bi: BotIntegration,
+    start_date: datetime.datetime | None = None,
+    end_date: datetime.datetime | None = None,
+) -> dict[str, typing.Any]:
+    msgs = Message.objects.filter(
+        conversation__bot_integration=bi,
+    ).exclude(
+        analysis_result={},
+    )
+    if start_date:
+        msgs = msgs.filter(created_at__gte=start_date)
+    if end_date:
+        msgs = msgs.filter(created_at__lte=end_date)
+    results = related_json_field_summary(
+        Message.objects,
+        "analysis_result",
+        qs=msgs,
+        query_param="--",
+        instance_id=bi.id,
+        max_keys=20,
+    )
+    return results
+
+
+def render_editor(
+    bi: BotIntegration,
+    results: dict[str, typing.Any] | None,
+    key: str,
+) -> list[dict[str, typing.Any]]:
+    from bots.models import BotIntegrationAnalysisChart
+
+    gui.write(
+        "Your bot's analysis script results can be visualized here. "
+        "Make sure your analysis prompts return a consistent JSON Schema for best results!"
+    )
+
+    # Load existing charts from DB
+    if key not in gui.session_state:
+        gui.session_state[key] = bi.analysis_charts.values(
+            "result_field", "graph_type", "data_selection"
+        )
+    # Allow editing in the UI
+    selected_graphs = list_view_editor(
+        add_btn_label="Add a Graph",
+        add_btn_type="tertiary",
+        key=key,
+        render_inputs=partial(render_inputs, results),
+    )
+
+    with gui.div(className="d-flex justify-content-end gap-3"):
+        if gui.session_state.get(key + ":save"):
+            BotIntegrationAnalysisChart.objects.filter(bot_integration=bi).delete()
+            for d in selected_graphs:
+                BotIntegrationAnalysisChart.objects.create(
+                    bot_integration=bi,
+                    result_field=d["result_field"],
+                    graph_type=d["graph_type"],
+                    data_selection=d["data_selection"],
+                )
+            # Optionally, reload the page or update the graphs
+            gui.success("Charts saved!")
+
+        gui.button(f"{icons.save} Save", type="primary", key=key + ":save")
+
+    return selected_graphs
+
+
+def render_graphs(
+    bi: BotIntegration,
+    results: dict[str, typing.Any] | None,
+    selected_graphs: list[dict[str, typing.Any]],
+):
     with gui.div(className="pb-5 pt-3"):
-        grid_layout(2, graphs, partial(render_graph_data, bi, results), separator=False)
-
-    gui.checkbox("🔄 Refresh every 10s", key="autorefresh")
-
-    with gui.expander("✏️ Edit"):
-        title = gui.text_area("##### Title", value=title)
-
-        gui.session_state.setdefault("selected_graphs", graphs)
-        selected_graphs = list_view_editor(
-            add_btn_label="Add a Graph",
-            key="selected_graphs",
-            render_inputs=partial(render_inputs, results),
+        grid_layout(
+            2,
+            selected_graphs,
+            partial(render_graph_data, bi=bi, results=results),
+            separator=False,
         )
 
-        with gui.center():
-            if gui.button("✅ Update"):
-                _on_press_update(title, selected_graphs)
 
-
-def render_inputs(results: dict, key: str, del_key: str, d: dict):
+def render_inputs(
+    results: dict[str, typing.Any], key: str, del_key: str, d: dict[str, typing.Any]
+):
     ocol1, ocol2 = gui.columns([11, 1], responsive=False)
     with ocol1:
         col1, col2, col3 = gui.columns(3)
@@ -90,14 +141,13 @@ def render_inputs(results: dict, key: str, del_key: str, d: dict):
         del_button(del_key)
 
     with col1:
-        d["key"] = gui.selectbox(
-            label="##### Key",
+        d["result_field"] = gui.selectbox(
+            label="##### JSON Field",
             options=results.keys(),
             key=f"{key}_key",
-            value=d.get("key"),
+            value=d.get("result_field"),
         )
     with col2:
-        col2.node.props["style"] = dict(paddingTop="0.45rem")
         d["graph_type"] = gui.selectbox(
             label="###### Graph Type",
             options=[g.value for g in GraphType],
@@ -106,7 +156,6 @@ def render_inputs(results: dict, key: str, del_key: str, d: dict):
             value=d.get("graph_type"),
         )
     with col3:
-        col3.node.props["style"] = dict(paddingTop="0.45rem")
         d["data_selection"] = gui.selectbox(
             label="###### Data Selection",
             options=[d.value for d in DataSelection],
@@ -130,94 +179,13 @@ def _autorefresh_script():
     )
 
 
-def render_title_breadcrumb_share(
+def render_graph_data(
+    graph_data: dict[str, typing.Any],
+    *,
     bi: BotIntegration,
-    current_url: str,
-    current_user: AppUser | None,
+    results: dict[str, typing.Any],
 ):
-    if bi.published_run_id:
-        run_title = bi.published_run.title
-        query_params = dict(example_id=bi.published_run.published_run_id)
-    else:
-        run_title = bi.saved_run.page_title  # this is mostly for backwards compat
-        query_params = dict(run_id=bi.saved_run.run_id, uid=bi.saved_run.uid)
-    with gui.div(className="d-flex justify-content-between mt-4"):
-        with gui.div(className="d-lg-flex d-block align-items-center"):
-            with gui.tag("div", className="me-3 mb-1 mb-lg-0 py-2 py-lg-0"):
-                with gui.breadcrumbs():
-                    metadata = VideoBotsPage.workflow.get_or_create_metadata()
-                    gui.breadcrumb_item(
-                        metadata.short_title,
-                        link_to=VideoBotsPage.app_url(),
-                        className="text-muted",
-                    )
-                    if not (bi.published_run_id and bi.published_run.is_root()):
-                        gui.breadcrumb_item(
-                            run_title,
-                            link_to=VideoBotsPage.app_url(**query_params),
-                            className="text-muted",
-                        )
-                    gui.breadcrumb_item(
-                        "Integrations",
-                        link_to=VideoBotsPage.app_url(
-                            **query_params,
-                            tab=RecipeTabs.integrations,
-                            path_params=dict(integration_id=bi.api_integration_id()),
-                        ),
-                    )
-
-            render_author_from_user(
-                bi.created_by,
-                show_as_link=current_user and VideoBotsPage.is_user_admin(current_user),
-            )
-
-        with gui.div(className="d-flex align-items-center"):
-            with gui.div(className="d-flex align-items-start right-action-icons"):
-                copy_to_clipboard_button(
-                    f'{icons.link} <span class="d-none d-lg-inline">Copy Link</span>',
-                    value=current_url,
-                    type="secondary",
-                    className="mb-0 ms-lg-2",
-                )
-
-
-def _on_press_update(title: str, selected_graphs: list[dict]):
-    if selected_graphs:
-        graphs_json = json.dumps(
-            [
-                dict(
-                    key=d["key"],
-                    graph_type=d["graph_type"],
-                    data_selection=d["data_selection"],
-                )
-                for d in selected_graphs
-            ]
-        )
-    else:
-        graphs_json = None
-    raise QueryParamsRedirectException(dict(title=title, graphs=graphs_json))
-
-
-@gui.cache_in_session_state
-def fetch_analysis_results(bi: BotIntegration) -> dict:
-    msgs = Message.objects.filter(
-        conversation__bot_integration=bi,
-    ).exclude(
-        analysis_result={},
-    )
-    results = related_json_field_summary(
-        Message.objects,
-        "analysis_result",
-        qs=msgs,
-        query_param="--",
-        instance_id=bi.id,
-        max_keys=20,
-    )
-    return results
-
-
-def render_graph_data(bi: BotIntegration, results: dict, graph_data: dict):
-    key = graph_data["key"]
+    key = graph_data["result_field"]
     gui.write(f"##### {key}")
 
     if graph_data["data_selection"] == DataSelection.last.value:
@@ -261,7 +229,7 @@ def render_graph_data(bi: BotIntegration, results: dict, graph_data: dict):
             render_radar_plot(values)
 
 
-def render_table_count(values):
+def render_table_count(values: list[list[typing.Any]]):
     if not values:
         gui.write("No analysis results found")
         return
@@ -273,55 +241,110 @@ def render_table_count(values):
     )
 
 
-def render_bar_chart(values):
+def render_bar_chart(values: list[list[typing.Any]]):
     import plotly.graph_objects as go
 
-    render_data_in_plotly(
+    labels = [result[0] for result in values]
+    short_labels = [truncate_label(label) for label in labels]
+    y_vals = [result[1] for result in values]
+    fig = make_figure(
         go.Bar(
-            x=[result[0] for result in values],
-            y=[result[1] for result in values],
-        ),
+            x=short_labels,
+            y=y_vals,
+            marker_color=COLOR_PALETTE,
+            customdata=labels,
+            hovertemplate="<b>%{customdata}</b><br>Count: %{y}<extra></extra>",
+        )
     )
+    fig.update_layout(
+        margin=dict(l=0, r=0, t=0, b=70),
+        xaxis=dict(tickangle=-35),
+    )
+    gui.plotly_chart(fig, config={"displayModeBar": False, "displaylogo": False})
 
 
-def render_pie_chart(values):
+def render_pie_chart(values: list[list[typing.Any]]):
     import plotly.graph_objects as go
 
-    render_data_in_plotly(
+    labels = [result[0] for result in values]
+    short_labels = [truncate_label(label) for label in labels]
+    vals = [result[1] for result in values]
+    pull = [0.05] + [0] * (len(vals) - 1)
+
+    fig = make_figure(
         go.Pie(
-            labels=[result[0] for result in values],
-            values=[result[1] for result in values],
-        ),
+            marker=dict(colors=COLOR_PALETTE),
+            labels=short_labels,
+            values=vals,
+            textinfo="percent+label",
+            textposition="inside",
+            insidetextorientation="auto",
+            pull=pull,
+            showlegend=False,
+            customdata=labels,
+            hovertemplate="<b>%{customdata}</b><br>Count: %{value}<br>Percent: %{percent}<extra></extra>",
+        )
     )
+    fig.update_layout(
+        margin=dict(l=0, r=0, t=0, b=0),
+    )
+    gui.plotly_chart(fig, config={"displayModeBar": False, "displaylogo": False})
 
 
-def render_radar_plot(values):
+def render_radar_plot(values: list[list[typing.Any]]):
     import plotly.graph_objects as go
 
     r = [result[1] for result in values]
-    theta = [result[0] for result in values]
+    theta_full = [result[0] for result in values]
+    theta = [truncate_label(label) for label in theta_full]
     if r and theta:
         r.append(r[0])
         theta.append(theta[0])
-    render_data_in_plotly(
-        go.Scatterpolar(r=r, theta=theta, fill="toself", marker=dict(color="#02b3a1")),
+        theta_full.append(theta_full[0])
+    # uniform fill/line base color
+    fill_color = "#4ECDC4"
+
+    fig = make_figure(
+        go.Scatterpolar(
+            r=r,
+            theta=theta,
+            fill="toself",
+            fillcolor=fill_color,
+            line=dict(color=fill_color),
+            marker=dict(color=COLOR_PALETTE, size=8),
+            customdata=theta_full,
+            hovertemplate="<b>%{customdata}</b><br>Count: %{r}<extra></extra>",
+        )
     )
-
-
-def render_data_in_plotly(*data):
-    import plotly.graph_objects as go
-
-    fig = go.Figure(
-        data=data,
-        layout=go.Layout(
-            template="plotly_white",
-            font=dict(size=16, family="basiercircle, sans-serif"),
-            polar=dict(
-                radialaxis=dict(angle=90, tickangle=90),
-                angularaxis=dict(rotation=25),
-            ),
-            margin=dict(l=40, r=40, t=40, b=40),
-            dragmode="pan",
-        ),
+    fig.update_layout(
+        margin=dict(l=70, r=70, t=70, b=70),
     )
     gui.plotly_chart(fig)
+
+
+def make_figure(*data: typing.Any) -> "go.Figure":
+    import plotly.graph_objects as go
+
+    fig = go.Figure(data=data)
+    fig.update_layout(
+        template="plotly_white",
+        font=dict(size=12, family="basiercircle,sans-serif"),
+        polar=dict(
+            radialaxis=dict(angle=90, tickangle=90),
+            angularaxis=dict(rotation=25),
+        ),
+        dragmode="pan",
+        autosize=True,
+        hovermode="x unified",
+    )
+    fig.update_xaxes(
+        spikemode="across",
+        spikedash="solid",
+        spikecolor="rgba(126, 87, 194, 0.25)",  # soft purple, semi-transparent
+        spikethickness=1,
+    )
+    return fig
+
+
+def truncate_label(text: str) -> str:
+    return truncate_text_words(text, 12, sep="…")
